Allow none value for colors in [default]
If foreground/background fields in [default] were
set to 'none', fbless crashed with TypeError.  If
the fields were missing, then the colors were taken
from defaults.py.  So, there were no chance to use
environment defaults.  Now, 'none' in defaults is
converted to -1, which says to curses that it should
use defult values.

      * main.py (MainWindow.init_color):
        If configure file had 'none' value for
	background or foreground fields, pass
	constant -1 to the curses.